refector: 야바위 애니매이션 코드 분리
