+++ src/entities/mini-game/yavarwee/index.ts
@@ -1,3 +1,7 @@
-export { default as YavarweeAnimation } from './ui/YavarweeAnimation';
 export { default as YavarweeButton } from './ui/YavarweeButton';
 export { default as Round } from './ui/Round';
+export { IdleOverlay } from './ui/IdleOverlay';
+export { RoundOverlay } from './ui/RoundOverlay';
+export { SelectingTimerProgress } from './ui/SelectingTimerProgress';
+export { StatusText } from './ui/StatusText';
+export { useCanvasAnimation } from './model/useCanvasAnimation';
