+++ src/shared/types/mini-game/yavarwee/index.ts
@@ -17,3 +17,33 @@ export interface YavarweeConfirmFormData {
 export interface YavarweeComfirmResponse {
   amount: number;
 }
+
+export interface CupState {
+  x: number;
+  y: number;
+  targetX: number;
+  targetY: number;
+  zIndex: number;
+  originalId: number;
+  animationProgress: number;
+  animationDuration: number;
+}
+
+export interface BallState {
+  x: number;
+  y: number;
+  visible: boolean;
+  scale: number;
+}
+
+export type GameState =
+  | 'idle'
+  | 'betting'
+  | 'showing'
+  | 'hiding'
+  | 'shuffling'
+  | 'selecting'
+  | 'result'
+  | 'round';
+
+export type Result = 'correct' | 'wrong' | null;
