+++ src/widgets/mini-game/yavarwee/index.ts
@@ -1,1 +1,2 @@
 export { default as RoundContainer } from './ui/RoundContainer';
+export { default as YavarweeAnimation } from './ui/YavarweeAnimation';
